Add/data/qualities/list (#99)
* Add datasets/qualities/list endpoint

* Add static test for qualities/list

* Fetch dataset qualities from the database

* Add test to verify behavior is same as PHP endpoint

Ideally we would want to test this against several database states
for example by deleting or adding new qualities. However, this
will not work without committing to the database. So we either need
to spin up a new database and php api for this test to be idempotent
or we lose idempotency. Neither seem worth the overhead.

* Add clarification on faithfulness and sketch V1-to-V2 changes